mindsdb/integrations/handlers/duckdb_faiss_handler/requirements.txt
duckdb>=0.9.0
faiss-cpu>=1.7.4
numpy>=1.21.0
pandas>=1.3.0
portalocker
